Moved image comparison functions to ImageBufAlgo (instead of being local to idiff)
